fix: refactor routes in flask, refactor InventoryList.jsx, change delete button in delete modal to primary, after deleting anything always display confiramtion, change label on the button in the inventory, add a unit to the polling frequency
