+++ backend/SC4SNMP_UI_backend/__init__.py
@@ -17,8 +17,12 @@ mongo_client = MongoClient(MONGO_URI)
 def create_app():
     app = Flask(__name__)
 
-    from SC4SNMP_UI_backend.ui_handling.routes import ui
-    app.register_blueprint(ui)
+    from SC4SNMP_UI_backend.profiles.routes import profiles_blueprint
+    from SC4SNMP_UI_backend.groups.routes import groups_blueprint
+    from SC4SNMP_UI_backend.inventory.routes import inventory_blueprint
+    app.register_blueprint(profiles_blueprint)
+    app.register_blueprint(groups_blueprint)
+    app.register_blueprint(inventory_blueprint)
     gunicorn_logger = logging.getLogger('gunicorn.error')
     app.logger.handlers = gunicorn_logger.handlers
     app.logger.setLevel(gunicorn_logger.level)
