+++ backend/SC4SNMP_UI_backend/common/helpers.py
